Add checks to clean XML reserved characters when used in page metadata (#3)
A user might want to use XML reserved characters in their page metadata, not realising that this would break the sitemap or RSS feed. This PR implements a cleaner function to ensure such characters are automatically replaced with escape codes at runtime.

This PR also adds a new unit test to make sure that the sanitisation actually works so that we don't break it down the line.