Make deployment workflow intelligently choose to deploy to main or staging (#5)
The way the deployment workflow worked originally is that it was hardcoded to deploy to a branch html, so if we wanted to test changes in staging we had to manually edit the workflow.

This PR adds a check to determine whether the branch we're working on is main or not. If it's main, deployments go to production. If it's not, deployments go to staging.